Generate a new dummy node for each test (#4)
Generate a new dummy node for each test

Previously the dummy node was being reused without using a deep copy. Some
tests were modifying the dictionary. This caused tests to fail if they were run
in same worker process.

The behaviour could be triggered when running the test suite with the serial flag
or when the number of tests was greater than the number of installed processors
(default behaviour is one worker process per cpu).

A new data structure is now created for each test. This fixes issue #3.