Ensure product name is consistent with introspection data
There was a discrepancy in the value of product name, as listed
in the introspection data, when compared with the one obtained via
dmidecode. This uses a utility method, get_system_vendor_info,
on the GenericHardwareManager to get the product name. This is
the same data as returned to the inspector, so it should
match.