Fix after ironic python agent py2 drop
Ironic Python Agent has dropped python 2 on the master branch. The last
release to support python 2 is Train. We currently hit a problem in
ironic-lib in our py27 tests:

ImportError: No module named ipaddress

This change uses the last version of IPA to support python 2, and the
Train upper constraints when testing python 2.